Reduce a file size using T-SQL
In SQL Server, when using Transact-SQL, reducing the size of a data file requires a different syntax to increasing it. To reduce a file size using T-SQL, use the DBCC SHRINKFILE command, and provide the name of the data file along with the size you’d like to reduce it to.